requirements.txt
kafka
cassandra-driver